DAS-2060: Migrate _open_datatree_zarr chunk.
